requirements.txt
sshtunnel
psycopg2-binary
reportlab
pandas
matplotlib
seaborn
nltk
